Need help on Mobile Responsive design
I need help on responsive design for this Email Campaign site that I am currently working on. The section called Find What's Right For You has a row of icons, 5 in total. I need these 5 icons in mobile to be staked on 2 different rows, the first row has 3 icons and the second row has 2 icons. Can someone provide / assist me how to do this for mobile version in responsive. Right now, all icons in mobile are stacked on top of each other in a long vertical lst.